+++ pioreactor/automations/dosing/__init__.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 
 from .chemostat import Chemostat
 from .fed_batch import FedBatch
-from .morbidostat import Morbidostat
 from .pid_morbidostat import PIDMorbidostat
 from .silent import Silent
 from .turbidostat import Turbidostat