+++ pioreactor/automations/temperature/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import annotations
 
-from .constant_duty_cycle import ConstantDutyCycle
 from .only_record_temperature import OnlyRecordTemperature
 from .thermostat import Thermostat